Added installing packages from Jupyter
I added a wonderful article by jakedvp. Essentially, in some cases it is
not advisable to install packages directly from Jupyter via a shell call
of pip install. There are several workarounds. If you happen to run into
this kind of trouble it'd be worthy to know his blog post.